collapse: nested-if → &&, bỏ cast thừa (clippy machine-safe) −28 LOC
PHÂN RÃ: thuần vế-I code-shape — `if a { if b {…} }` (không else) ≡
`if a && b {…}` (short-circuit cùng ngữ nghĩa); + cast same-type thừa,
`return` thừa, `repeat().take()`→`repeat_n`, `map_or` rút gọn. KHÔNG đổi
control-flow/thuật toán, chỉ nén cú pháp. Áp bằng `cargo clippy --fix`
(chỉ rewrite rustc-verified machine-applicable).

CHỨNG CỨ 0-regression (đo-trước-khi-nói, cả 2 path):
- asm-diff OLD(40e7e19)↔NEW toàn corpus: csmith 300 IDENTICAL,
  torture 1554 IDENTICAL, 0 DIFFER (--ir path byte-identical)
- base (cases+ext) PASS · torture box 1377/0 FAIL (AST-walk path)
- cargo test 29/29

src 11222 → 11194 LOC.

+++ src/codegen/arm64_elf.rs
@@ -267,7 +267,7 @@ fn emit_params(g: &mut Cg, f: &crate::ast::Func) {
             }
         }
     }
-    g.va = (gp.min(8) as u32, fp.min(8) as u32, boff, f.frame);
+    g.va = (gp.min(8), fp.min(8), boff, f.frame);
 }
